requirements.txt
langchain
Numpy
Pandas
Streamlit
langchain_community
python-dotenv
pypdf
bs4
wikipedia
langchain-text-splitters
ipykernel
langchain-openai
chromadb
langchain-ollama
sentence-transformers
langchain-huggingface
faiss-cpu
langchain_chroma